submission/requirements.txt
ultralytics>=8.0.0  # YOLOv8
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
tqdm>=4.65.0
scikit-learn>=1.3.0
transformers>=4.30.0  # For CLIP
ftfy>=6.1.1
regex>=2023.6.3
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0 